requirements.txt
fdb
python-telegram-bot
python-telegram-bot[job-queue]
psycopg2
python-dotenv